+++ dalla-stats.py
@@ -7,66 +7,77 @@ from os import listdir
 from os.path import isfile, join
 import csv
 import datetime
+import logging
+import sys
 
 def main():
     parser = argparse.ArgumentParser()
 
-    parser.add_argument("-u", "--username", help="the router admin username")
-    parser.add_argument("-p", "--password", help="the router admin password")
+    parser.add_argument("-u", "--username", default='', help="the router admin username")
+    parser.add_argument("-p", "--password", default='', help="the router admin password")
     parser.add_argument("-i", "--interval", type=int, default=0, help="the interval in seconds to update the statistics.")
-    parser.add_argument("-d", "--working-directory", default='.', help="directory to save logs")
+    parser.add_argument("-d", "--log-directory", default='logs', help="directory to save logs")
     parser.add_argument("-l", "--enable-logging", default=False, action='store_true', help="Log statistics?")
     parser.add_argument('-v', '--version', action='version', version='%(prog)s v0.0.2')
 
     args = parser.parse_args()
 
+    if (args.username == '' or args.password == ''):
+        print('[ERROR] Please supply username and password')
+        exit()
+
+    logDir = args.log_directory
+
+    session = initSession(args.username, args.password)
+
+    userMap = loadUserMap(logDir)
+
+    oldStats = loadDeviceSummary(logDir)
+
     try:
-        os.mkdir(args.working_directory)
+        os.mkdir(logDir)
     except OSError:
         i = 5
 
-    summaryPath = args.working_directory + '/devices.csv'
-    devicePath = args.working_directory + '/devices/'
-    userDir = args.working_directory + '/users/'
-    totalPath = args.working_directory + '/total.csv'
-    mapPath = args.working_directory + '/user-map.csv'
-
-    session = requests.session()
-    initSession(args.username, args.password, session)
-
-    userMap = loadUserMap(mapPath)
-
-    oldStats = loadDeviceSummary(summaryPath)
+    print('[INFO] Starting...')
 
     while (True):
 
         deviceStats = getDeviceRecords(session)
-        delta = calculateDeviceDeltas(oldStats, deviceStats)
 
-        mergeDevices(oldStats, delta)
-        saveDeviceSummary(delta, summaryPath)
+        if (len(deviceStats) != 0):
+            delta = calculateDeviceDeltas(oldStats, deviceStats)
+            mergeDevices(oldStats, delta)
 
-        if (args.enable_logging):
-            logDeviceStats(delta, devicePath)
+            saveDeviceSummary(delta, logDir)
 
-            userStats = getUserStats(delta, userMap)
-            logUserStats(userStats, userDir)
+            if (args.enable_logging):
+                logDeviceStats(delta, logDir)
 
-            total = getTotalStats(userStats)
-            logTotalStats(total, totalPath)
+                userStats = getUserStats(delta, userMap)
+                logUserStats(userStats, logDir)
+
+                total = getTotalStats(userStats)
+                logTotalStats(total, logDir)
+
+            oldStats = delta
 
         if ( args.interval == 0):
             break
 
-        oldStats = delta
-
         time.sleep(args.interval)
 
-def saveDeviceSummary(deviceStatsArray, summaryFile):
+def saveDeviceSummary(deviceStatsArray, logDir):
     """Save the given dict array to file
     """
 
-    summaryFile = open(summaryFile, 'w')
+    try:
+        os.mkdir(logDir)
+    except OSError:
+        i = 5
+
+    summaryFile = open(logDir + '/deviceSummary.csv', 'w')
+
 
     summaryFile.write('MAC Address, IP Address, Time,Total Bytes, Delta, On-Peak, Off-Peak\n')
 
@@ -78,7 +89,7 @@ def saveDeviceSummary(deviceStatsArray, summaryFile):
     summaryFile.close()
 
 
-def loadDeviceSummary(summaryFile):
+def loadDeviceSummary(logDir):
     """Load the device summary into a dict array
     """
 
@@ -87,10 +98,10 @@ def loadDeviceSummary(summaryFile):
     """
 
     deviceStats = []
-    if (os.path.isfile(summaryFile) == False):
+    if (os.path.isfile(logDir + '/deviceSummary.csv') == False):
         return []
 
-    summaryFile = open(summaryFile, 'r')
+    summaryFile = open(logDir + '/deviceSummary.csv', 'r')
     reader = csv.reader(summaryFile, delimiter=',', skipinitialspace=True)
 
     for row in reader:
@@ -174,8 +185,29 @@ def getDeviceRecords(session):
     session.headers.update({'Referer': 'http://192.168.1.1/mainFrame.htm'})
     data ='[STAT_CFG#0,0,0,0,0,0#0,0,0,0,0,0]0,0\r\n[STAT_ENTRY#0,0,0,0,0,0#0,0,0,0,0,0]1,0\r\n'
 
-    r = session.post(url=url, data=data)
+    try:
+        r = session.post(url=url, data=data, timeout=1)
+    except requests.ConnectionError:
+        print('[ERROR] Network unreachable!')
+        return {}
+    except:
+        print('[ERROR] Unexpected error: ', sys.exc_info()[0])
+        return {}
+
+
+
     rawStats = r.text
+
+    error = rawStats.split('\n')
+
+    if (error[-1] != '[error]0'):
+        print('[ERROR] Failed to get device records from router!')
+        if (r.text == '<html><head><title>500 Internal Server Error</title></head><body><center><h1>500 Internal Server Error</h1></center></body></html>'):
+            print('\t Another admin has logged in!')
+        else:
+            print('\t' + r.text)
+
+        return []
 
     dictArray = []
     tmpDict = {}
@@ -210,6 +242,8 @@ def getDeviceRecords(session):
 
     # Manipulate dict array to get what we need
     init = initDevices(dictArray, timeKey)
+
+    logout(session)
 
     return init
 
@@ -259,6 +293,7 @@ def calculateDeviceDeltas(oldDeviceDeltas, currentDeviceRecords):
                         newDeviceDict['Delta'] = newDeviceDict['Total Bytes']
 
                         print(oldDeviceDict)
+                        print('')
 
                     classifyDelta(newDeviceDict)
 
@@ -270,6 +305,7 @@ def calculateDeviceDeltas(oldDeviceDeltas, currentDeviceRecords):
             classifyDelta(newDeviceDict)
 
             print(newDeviceDict)
+            print('')
 
     return localCurrent
 
@@ -282,7 +318,7 @@ def logDeviceStats(statsDictArray, prefix):
     """
 
     try:
-        os.mkdir(prefix)
+        os.mkdir(prefix + '/devices/')
     except OSError:
         i = 5
 
@@ -290,7 +326,7 @@ def logDeviceStats(statsDictArray, prefix):
         # Generate file name
         mac = statsDict['MAC Address'].replace(':', '-')
         ip = statsDict['IP Address']
-        fileName = str(prefix + mac + '_' + ip + '.csv')
+        fileName = str(prefix + '/devices/' + mac + '_' + ip + '.csv')
 
         # csv fields
         timeKey = statsDict['Time']
@@ -343,7 +379,7 @@ def getUserStats(deviceStatsArray, userMap):
         #     print('[INFO] Time key mismatch! (getUserStats)')
 
         # Use usermap to determine to who this mac beints
-        if (userMap.has_key(mac)):
+        if (mac in userMap):
 
             # Have we seen this user before?
             found = False
@@ -376,18 +412,18 @@ def getUserStats(deviceStatsArray, userMap):
 
     return userStatsArray
 
-def logUserStats(userStatsArray, userLogPath):
+def logUserStats(userStatsArray, logDir):
     """
     Append the user data to their csv files
     """
 
     try:
-        os.mkdir(userLogPath)
+        os.mkdir(logDir + '/users/')
     except OSError:
         i = 5
 
     for user in userStatsArray:
-        fileName = userLogPath + user['Name'] + '.csv'
+        fileName = logDir + '/users/' + user['Name'] + '.csv'
         header = False
 
         if (os.path.isfile(fileName) == False):
@@ -427,18 +463,24 @@ def getTotalStats(userStats):
 
     return total
 
-def logTotalStats(totalStats, totalLogFile):
+def logTotalStats(totalStats, logDir):
     """
     Append the total to their csv files
     """
 
+    fileName = logDir + '/total.csv'
+
+    try:
+        os.mkdir(logDir)
+    except OSError:
+        i = 5
+
     header = False
 
-    if (os.path.isfile(totalLogFile) == False):
+    if (os.path.isfile(fileName) == False):
         header = True
 
-    totalcsv = open(totalLogFile, 'a')
-
+    totalcsv = open(fileName, 'a')
     if (header):
         totalcsv.write('Time, Total Bytes, Delta, On-Peak, Off-Peak\n')
 
@@ -457,7 +499,9 @@ def getMacFromFileName(name):
 
     return result
 
-def initSession(username, password, session):
+def initSession(username, password):
+    session = requests.session()
+
     raw = username + ':' + password
     encoded = base64.b64encode(raw.encode('utf-8'))
 
@@ -475,6 +519,8 @@ def initSession(username, password, session):
         'Referer': 'http://192.168.1.1/',
         'Connection': 'keep-alive'
     }
+
+    return session
 
 def decStrToIpStr(dec):
     binStr = bin(int(dec))
@@ -501,18 +547,42 @@ def loadUserMap(path):
     MAC, User
     """
 
+    fileName = path + '/user-map.csv'
+
     userMap = {}
 
-    if (os.path.isfile(path) == False):
+    if (os.path.isfile(fileName) == False):
         return userMap
 
-    userMapFile = open(path)
+    userMapFile = open(fileName)
     reader = csv.reader(userMapFile, delimiter=',', skipinitialspace=True)
 
     for row in reader:
         if (reader.line_num != 1):
             userMap[row[1]] = row[0]
 
+    userMapFile.close()
+    
     return userMap
+
+def logout(session):
+    # Configure page specific headers
+    url = 'http://192.168.1.1/cgi?8'
+    session.headers.update({'Referer': 'http://192.168.1.1/MenuRpm.htm'})
+    data ='[/cgi/logout#0,0,0,0,0,0#0,0,0,0,0,0]0,0\r\n'
+
+    try:
+        r = session.post(url=url, data=data)
+    except:
+        print(str(datetime.datetime.now()) + ' [ERROR] Unexpected error: ', sys.exc_info()[0])
+        return
+
+
+    if (r.text != '[cgi]0\n[error]0'):
+        print('[ERROR] Logout failed:')
+        if (r.text == '<html><head><title>500 Internal Server Error</title></head><body><center><h1>500 Internal Server Error</h1></center></body></html>'):
+            print('\t Another admin has logged in!')
+        else:
+            print('\t' + r.text)
 
 main()